+++ Modules/Microsoft365DSC/DSCResources/MSFT_O365OrgSettings/MSFT_O365OrgSettings.psm1
@@ -156,6 +156,14 @@ function Get-TargetResource
     $ConnectionModeTasks = New-M365DSCConnection -Workload 'Tasks' `
         -InboundParameters $PSBoundParameters
 
+    # Workaround for issue when if connected to S+C prior to calling cmdlet, an error about an invalid token is thrown.
+    # If connected to S+C, then we need to re-initialize the connection to EXO.
+    if ($Global:MSCloudLoginConnectionProfile.SecurityComplianceCenter.Connected -and `
+        $Global:MSCloudLoginConnectionProfile.ExchangeOnline.Connected)
+    {
+        $Global:MSCloudLoginConnectionProfile.ExchangeOnline.Disconnect()
+        $Global:MSCloudLoginConnectionProfile.SecurityComplianceCenter.Connected = $false
+    }
     $ConnectionMode = New-M365DSCConnection -Workload 'ExchangeOnline' `
         -InboundParameters $PSBoundParameters
 
@@ -210,7 +218,7 @@ function Get-TargetResource
             }
         }
 
-        # Viva Insightss settings
+        # Viva Insights settings
         $currentVivaInsightsSettings = Get-DefaultTenantMyAnalyticsFeatureConfig
         $MRODeviceManagerService = 'ebe0c285-db95-403f-a1a3-a793bd6d7767'
         try